feat(kscli): 新增 Knowledge Studio CLI 轻量级 RAG 命令行工具
- 用于阿里云 Model Studio 的知识库检索，支持 RAG（检索增强生成）场景
- 提供配置查看与设置、知识库检索、自更新功能
- 替换原 rag 子包，移除 rag 相关代码及配置
- 新增独立 package，包含完整的构建、启动和发布配置
- 添加详细的中英文 README 文档说明安装、使用与认证方式
- 配置 TypeScript 和 Vite 构建支持，确保开发体验和构建质量
- 更新根 package.json 脚本，将 rag dev 命令替换为 kscli dev
- 新增 Git 忽略文件，排除日志、构建输出等无关文件
