+++ package.json
@@ -19,7 +19,7 @@
     "sync:skill-assets": "pnpm --filter bailian-cli-core run build && pnpm --filter bailian-cli run generate:reference && pnpm --filter bailian-cli run sync:skill-version",
     "dev": "pnpm -F bailian-cli-core dev",
     "bl": "pnpm -F bailian-cli dev",
-    "rag": "pnpm -F bailian-cli-rag dev",
+    "kscli": "pnpm -F knowledge-studio-cli dev",
     "test": "vp test",
     "release:check": "node tools/release/check.mjs",
     "wiki:crawl": "node tools/wiki-crawler/index.mjs",
