+++ packages/kscli/tsconfig.json
